+++ src/weather/mod.rs
@@ -2,11 +2,13 @@ use crate::config::Config;
 use crate::errors::RustormyError;
 use crate::models::{Location, Weather};
 use anyhow::Result;
-pub use open_meteo::OpenMeteoProvider;
-pub use open_weather_map::OpenWeatherMapProvider;
-pub use provider::WeatherProvider;
+use enum_dispatch::enum_dispatch;
+pub use open_meteo::OpenMeteo;
+pub use open_weather_map::OpenWeatherMap;
+pub use provider::GetWeatherProvider;
 
 #[async_trait::async_trait]
+#[enum_dispatch]
 pub trait GetWeather {
     async fn get_weather(&self, config: &Config) -> Result<Weather, RustormyError>;
     async fn lookup_city(&self, city: &str, config: &Config) -> Result<Location, RustormyError>;